Module Kqueue.Event_list

type t

Kqueue.Event_list.t represents a list of kevents that can be used to populate either the changelist or the eventlist for the kevent syscall. The intended usage is to allocate a list of events using create, and then retrieve events at a particular index using get idx, and then using the Kqueue.Event_list.Event module to read/write the varios fields that belong to a kevent.

val null : t

val create : int -> t

module Event : sig ... end

val get : t -> int -> Event.t