Module Kqueue.Flag

type t

Kqueue.Flag.t is a set of flags that are used to indicate which actions should be performed for an event.

val equal : t -> t -> bool

val (=) : t -> t -> bool

val pp : Stdlib.Format.formatter -> t -> unit

val (+) : t -> t -> t

val intersect : t -> t -> bool

val receipt : t

val add : t

val enable : t

val disable : t

val delete : t

val oneshot : t

val clear : t

val eof : t

val error : t