Module Kqueue.Kevent

Kevent.t is used to represent an input/output event that can be registered with the kernel. A kevent consists of:

ident: an identifier for an event.

filter: filter is used to determine if an event has met the pre-condition for it to be made available for retrieval by the user.

flags: flags are used to specify actions that need to be performed on an event.

fflags: filter specific flags.

data: filter specific data.

udata: User-defined data that gets passed through unchanged.

type t = Bindings.Kevent.t

val make : ident:Ctypes.Uintptr.t -> filter:int -> flags:Unsigned.uint16 -> fflags:Unsigned.uint32 -> data:Ctypes.Intptr.t -> udata:Ctypes.Uintptr.t -> t

val ident : t -> Ctypes.Uintptr.t

val filter : t -> int

val flags : t -> Unsigned.uint16

val fflags : t -> Unsigned.uint32

val data : t -> Ctypes.Intptr.t

val udata : t -> Ctypes.Uintptr.t