Module Kqueue.Note

type t

Kqueue.Note.t represents the kqueue filter specific flags. These are used in combination with the data field to augment the behavior of kqueue event filters.

Consult the kqueue manpages to see what the various Note values represent.

FreeBSD

OpenBSD

val equal : t -> t -> bool

val (=) : t -> t -> bool

val empty : t

val seconds : t

val useconds : t

val nseconds : t

val lowat : t

val oob : t

val delete : t

val write : t

val extend : t

val attrib : t

val link : t

val rename : t

val revoke : t

val exit : t

val fork : t

val exec : t

val signal : t

val ffnop : t

val ffand : t

val ffor : t

val ffcopy : t

val ffctrlmask : t

val fflagsmask : t

val trigger : t