Module Kqueue.Timeout

type t

Kqueue.Timeout.t is the timespan in nanoseconds that is used to represent the maximum amount of time the kevent call should wait for an event.

val never : t

never results in the kevent call to wait indefinitely for a new event.

Note: Unlike Unix.select, if the user calls kevent with an empty eventlist, the kevent call returns immediately event when a timeout value of never is used.

val immediate : t

immediate effects a poll, and the kevent call will return immediately, without waiting for new events if none are available at the time of the call.

val of_ns : int64 -> t

of_ns creates a new timeout from nanoseconds.