Module Kqueue.Util

val file_descr_to_int : Unix.file_descr -> int

file_descr_to_int will convert a Unix.file_descr to an integer. This is intended to be used to create an kevent ident when registering a file descriptor based kevent.

val file_descr_of_int : int -> Unix.file_descr

file_descr_to_int will convert an integer to a Unix.file_descr.