Module Kqueue

Kqueue provides bindings to the Kqueue event notification interface available on many BSD based systems and macOS.

module Bindings : module type of sig ... end

module Timespec : sig ... end

Timespec is used to set the maximum interval a kevent syscall will wait for an event.

module Kevent : sig ... end

Kevent.t is used to represent an input/output event that can be registered with the kernel. A kevent consists of:

type error = [

| `ENOMEM

| `EMFILE

| `ENFILE

| `EACCES

| `EFAULT

| `EBADF

| `EINTR

| `EINVAL

| `ENOENT

| `ESRCH

| `UNKNOWN of int

]

val error_message : error -> string

error_message returns a description of a kevent error code.

type t

val kqueue : unit -> (t, [> error ]) Stdlib.result

kqueue creates a new kernel event queue.

val kqueue_exn : unit -> t

val kevent : ?⁠timeout:Timespec.t -> t -> changelist:Kevent.t Ctypes_static.carray -> eventlist:Kevent.t Ctypes_static.carray -> (int, [> error ]) Stdlib.result

kevent is used to register events with kqueue, and fetch events that are available to read from the kqueue. The changelist array contains the new events that need to be registered. The items in the changelist are processed before any new items are read from a kqueue. If eventlist is non empty, the kevent call will block indefinitely till there are n events available to read, n being the length of the eventlist array. The optional timeout parameter can be used instead to specify the maximum interval that kevent should wait for before returning.

val kevent_exn : ?⁠timeout:Timespec.t -> t -> changelist:Kevent.t Ctypes_static.carray -> eventlist:Kevent.t Ctypes_static.carray -> int